updated mapgen to use src/random
