+++ src/mapgen/mapgen.c
@@ -31,9 +31,7 @@
 #include "ilist_test.h"
 #include "vectors/roads.h"
 #include "vectors/tiles.h"
-#include "sfc32_test.h"
-#include "random.h"
-#include "random_test.h"
+#include "random/random.h"
 
 #define        SEED_FILE    "randseed"
 
@@ -265,16 +263,6 @@ int main(int argc, char **argv) {
             case 't':
                 if (strcmp(optarg, "ilist") == 0) { // test ilist
                     if (0 != test_ilist()) {
-                        printf("test %s: failed\n", optarg);
-                        exit(2);
-                    }
-                } else if (strcmp(optarg, "prng") == 0) { // test prng
-                    if (0 != test_prng()) {
-                        printf("test %s: failed\n", optarg);
-                        exit(2);
-                    }
-                } else if (strcmp(optarg, "random") == 0) { // test random number function
-                    if (0 != test_random()) {
                         printf("test %s: failed\n", optarg);
                         exit(2);
                     }
