--- src/mapgen/random_test.h
@@ -1,11 +1,0 @@
-// olytag - Olympia: The Age of Gods
-//
-// Copyright (c) 2022 by the OlyTag authors.
-// Please see the LICENSE file in the root directory of this repository for further information.
-
-#ifndef OLYTAG_RANDOM_TEST_H
-#define OLYTAG_RANDOM_TEST_H
-
-int test_random(void);
-
-#endif //OLYTAG_RANDOM_TEST_H
